Give Barachians LOS 8 (gammafunk)
The general feeling was that Barachians were a bit bland, so let's
add a new mechanic to interact with their existing Hop and slow
movement. They now have +1 LOS, making considerably more spaces
visible than for normal races. This is generally considered Bad;
in exchange, thus, I'm raising their attributes somewhat, and
also raising most of their aptitudes by +1.

(Looking at it now, I suspect the aptitude changes might be
excessive; maybe revert that part and increase Hop's range
instead?)

As part of the change, the race is rethemed: instead of being
barbarians from the arctic south, they're now rebellious former
servants of the good gods, which naturally repel shadows. (The
opposite of Nightstalker, more or less.) Accordingly, they now have
poor Invocations and Necromancy apts.

It could be worth making the Robe of Night and the Darkness spell
not work for them, for theme reasons, but that's a long and
slippery slope. (Should Umbra not work? Should they be banned from
Dith...?)